% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caihelp-package.r
\docType{package}
\name{caihelp-package}
\alias{caihelp-package}
\title{A package to facilitate code generation to surface a C API}
\description{
A package to facilitate code generation to surface a C API
}
\details{
\tabular{ll}{
Package: \tab caihelp\cr
Type: \tab Package\cr
Version: \tab 0.7 \cr
Date: \tab 2021-01-04 \cr
Release Notes: \tab Refactores out of package capigen \cr
License: \tab GPL 2\cr
}

\tabular{lll}{
Version \tab Date \tab Notes \cr
0.6.1 \tab 2018-02-27 \tab Fix: codegen was catering for references, which is not a C "thing" and should not be in C APIs.\cr
}
}
\author{
Jean-Michel Perraud \email{jean-michel.perraud_at_csiro.au}
}
\keyword{Matlab}
