% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-functions.r
\name{find_env_var_as_path}
\alias{find_env_var_as_path}
\title{Find the path specified by the env var SwiftSrcPath}
\usage{
find_env_var_as_path(env_var_name)
}
\description{
Find the path specified by the env var SwiftSrcPath
}
