% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proto.r
\name{get_api_functions}
\alias{get_api_functions}
\title{Extract C API functions from a header file}
\usage{
get_api_functions(header_file, api_marker = "SWIFT_API")
}
\arguments{
\item{header_file}{location of the C API header file}

\item{api_marker}{string that, if found first in a line, identifies the line as a function to wrap}
}
\value{
character vector, the API functions
}
\description{
Extract C API functions from a header file
}
