% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-functions.r
\name{purlAll}
\alias{purlAll}
\title{Extract the R code from R Markdown}
\usage{
purlAll(outDir, inDir, pattern = "\\\\.Rmd$")
}
\arguments{
\item{outDir}{the output directory}

\item{inDir}{The directory with the .Rmd files}

\item{pattern}{a file pattern to look for}
}
\description{
Extract the R code from R Markdown
}
