% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-roxygen.r
\name{roxy_ncSwift}
\alias{roxy_ncSwift}
\title{roxygenize ncSwift R package}
\usage{
roxy_ncSwift(
  swiftSrcPath = "",
  pathRelativeToSwift = "../netcdf-tools/R/pkgs/ncSwift"
)
}
\arguments{
\item{swiftSrcPath}{the root of swift src codebase. If empty string, find_env_SwiftSrcPath() is used.}

\item{pathRelativeToSwift}{Location of the package to roxygenize}
}
\description{
roxygenize ncSwift R package
}
