% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-roxygen.r
\name{roxy_pkg}
\alias{roxy_pkg}
\title{roxygenize R package}
\usage{
roxy_pkg(swiftSrcPath = "", pathRelativeToSwift = "R/pkgs/swift")
}
\arguments{
\item{swiftSrcPath}{the root of swift src codebase. If empty string, find_env_SwiftSrcPath() is used.}

\item{pathRelativeToSwift}{Location of the package to roxygenize}
}
\description{
roxygenize R package
}
