% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-roxygen.r
\name{roxy_rpp}
\alias{roxy_rpp}
\title{roxygenize rpp}
\usage{
roxy_rpp(
  swiftSrcPath = "",
  pathRelativeToSwift = "../rpp-cpp/bindings/R/pkgs/rpp"
)
}
\arguments{
\item{swiftSrcPath}{the root of swift src codebase. If empty string, find_env_SwiftSrcPath() is used.}

\item{pathRelativeToSwift}{Location of the package to roxygenize}
}
\description{
roxygenize rpp
}
