% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-functions.r
\name{str_replace_file}
\alias{str_replace_file}
\title{Apply stringr::str_replace_all to all lines of a file}
\usage{
str_replace_file(filename, pattern, replacement)
}
\arguments{
\item{filename}{the file to parse and modify}

\item{pattern}{pattern}

\item{replacement}{replacement}
}
\description{
Apply stringr::str_replace_all to all lines of a file
}
