% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd-functions.r
\name{swiftDotnetBinDir}
\alias{swiftDotnetBinDir}
\title{Gets the location of the CLR binaries of SWIFT}
\usage{
swiftDotnetBinDir(srcDir = find_env_SwiftSrcPath(), buildConfig = "Debug")
}
\arguments{
\item{srcDir}{root of the SWIFT codebase}

\item{buildConfig}{The build configuration to use; 'Debug' or 'Release'}
}
\description{
Gets the location of the CLR binaries of SWIFT
}
